requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
Naked==0.1.32
pycryptodome==3.17
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
urllib3==1.26.15
